Server/requirements.txt
beautifulsoup4==4.13.3
cryptography==44.0.2
DBUtils==3.1.0
Flask==3.1.0
pycryptodome==3.21.0
PyMySQL==1.1.1
pyquery==2.0.1
Requests==2.32.3
